+++ Network/API/Telegram/Bot/Object/Special/Name.hs
@@ -1,4 +1,4 @@
-module Network.API.Telegram.Bot.Object.Name
+module Network.API.Telegram.Bot.Object.Special.Name
 	(Name (..), First (..), Last (..), Nick (..)) where
 
 import "aeson" Data.Aeson (FromJSON (parseJSON)